update service layer to move to the spring

+++ src/main/java/com/store/services/impl/BasketServiceImpl.java
@@ -1,18 +1,19 @@
 package com.store.services.impl;
 
 import com.store.entities.Order;
+import com.store.exceptions.ResourceNotFoundException;
 import com.store.repositories.BasketRepository;
 import com.store.services.BasketService;
-
+import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
+@Service
+@AllArgsConstructor
 public class BasketServiceImpl implements BasketService {
     private final BasketRepository repository;
-
-    public BasketServiceImpl(BasketRepository repository) {
-        this.repository = repository;
-    }
 
     @Override
     public List<Order> getAllOrdersByUserId(int id) {
@@ -26,12 +27,22 @@ public class BasketServiceImpl implements BasketService {
 
     @Override
     public boolean delete(Order order) {
-        return repository.delete(order);
+        Optional<Order> entity = repository.getById(order.getId());
+        if(entity.isPresent()){
+            return repository.delete(order);
+        }
+
+        throw new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Order wasn't found");
     }
 
     @Override
     public Optional<Order> getById(int id) {
-        return repository.getById(id);
+        Optional<Order> order = repository.getById(id);
+
+        if (order.isPresent()) {
+            return order;
+        }
+        throw new ResourceNotFoundException(HttpStatus.NOT_FOUND, "Order wasn't found");
     }
 
     @Override
